requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
platformdirs==3.2.0
pooch==1.7.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.3.0
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pyzmq==23.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.11.0
seaborn==0.12.2
setuptools==65.6.3
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.21
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.6
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.15.0
zipp==3.15.0
